moved Navigation into common/Navigation
