+++ src/Navigation/Search/SearchPage/SearchedPeople.js
@@ -4,8 +4,8 @@ import { Section, SectionTitle } from "../../../common/Section/Section";
 import { SmallListWrapper, StyledLink } from "../../../common/Tile/styled";
 import { ListTileSmall, } from "../../../common/Tile";
 import Pagination from "../../../common/Pagination";
-import { selectCurrnetPage, selectTotalPages } from "../searchSlice";
-import { goToFirstSearchPage, goToLastSearchPage, incrementPage, decrementPage } from "../searchSlice";
+import { selectCurrnetPage, selectTotalPages } from "../../../common/Navigation/Search/searchSlice";
+import { goToFirstSearchPage, goToLastSearchPage, incrementPage, decrementPage } from "../../../common/Navigation/Search/searchSlice";
 import AnimatedPage from "../../../common/AnimatedPage";
