+++ src/app/store.js
@@ -4,7 +4,7 @@ import rootSaga from './rootSaga';
 import movieListReducer from '../features/movies/MovieList/movieListSlice';
 import movieReducer from '../features/movies/MovieDetails/movieSlice';
 import peopleReducer from '../features/people/PeopleList/peopleSlice';
-import searchReducer from "../Navigation/Search/searchSlice";
+import searchReducer from "../common/Navigation/Search/searchSlice";
 import peopleDetailsReducer from "../features/people/PeopleDetails/peopleDetailsSlice";
 
 const sagaMiddleware = createSagaMiddleware();
