+++ src/features/people/PeopleDetails/index.js
@@ -16,7 +16,7 @@ import { Main } from "../../../common/Main/Main";
 import Error from "../../../common/Error";
 import Loading from "../../../common/Loading";
 import AnimatedPage from "../../../common/AnimatedPage";
-import { goToFirstSearchPage, selectInputValue, setInputValue, selectData } from "../../../Navigation/Search/searchSlice";
+import { goToFirstSearchPage, selectInputValue, setInputValue, selectData } from "../../../common/Navigation/Search/searchSlice";
 import { SearchPage } from "../../../Navigation/Search/SearchPage";
